Moved post author-relevant components to 'author' directory to organize files better
