+++ frontend/src/features/posts/components/author/PostAuthorThumbnail.jsx
@@ -1,7 +1,7 @@
 import { Suspense, useState } from "react";
 import PostAuthorDropdown from "./PostAuthorDropdown";
-import LoadingSvg from "../../../assets/LoadingSvg";
-import ErrorSvg from "../assets/ErrorSvg";
+import LoadingSvg from "../../../../assets/LoadingSvg";
+import ErrorSvg from "../../assets/ErrorSvg";
 const PostAuthorThumbnail = ({ profile, status }) => {
   const [showDropdown, setShowDropdown] = useState(false);
   return (
